MAX30100 - TIVAC
The objective of the project is to be able to read the data received from the MAX30100 sensor via TCP/IP server. he command sequences in the sockets created for these operations are as follows:
The data command to be received from the sensor is "READ HEARTBEAT", the real time command to be received with NTP is "TIME" and the "SHUTDOWN" commands to terminate the socket task.